# jaxrs/src/main/resources/schema/swagger.yaml
openapi: 3.0.1
info:
  title: Fuse TryIt - Jaxrs API
  version: 1.0.0
...
paths:
...
  /user/{userId}/model/{modelId}/{runtime}/build:
    get:
      summary: Schedule the build process for the model.
      operationId: buildModel
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      - name: modelId
        in: path
        required: true
        schema:
          type: string
      - name: runtime
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "[OK] Successfully scheduled the model build."
        "401":
          description: "[Unauthorized] If the provided access token was not valid."
        "404":
          description: "[Not Found] The model for the given id could not be found."
  /user/{userId}/model/{modelId}/{runtime}/download:
    get:
      summary: Download the build target
      operationId: downloadBuildTarget
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      - name: modelId
        in: path
        required: true
        schema:
          type: string
      - name: runtime
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "[OK] Found the requested build target."
          content:
            application/octet-stream: {}
        "401":
          description: "[Unauthorized] If the provided credentials were not valid."
        "404":
          description: "[Not Found] The target file was not found."
  /user/{userId}/model/{modelId}/{runtime}/status:
    get:
      summary: Get the current build status
      operationId: getBuildStatus
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      - name: modelId
        in: path
        required: true
        schema:
          type: string
      - name: runtime
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "[OK] Found the status for the requested project."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MavenBuildHandle'
        "401":
          description: "[Unauthorized] If the provided credentials were not valid."
        "404":
          description: "[Not Found] The project for the given id was not found."
...
  /users:
    put:
      summary: Register a new user with Keycloak
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        "201":
          description: "[Created] User was successfully created to Keycloak."
        "409":
          description: "[Conflict] If the user already exists in Keycloak."
  /users/login:
    post:
      summary: Login the user with Keycloak
      operationId: userLogin
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              required:
              - password
              - username
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: "[OK] Successful user login."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserTokens'
        "401":
          description: "[Unauthorized] If the provided credentials were not valid."
components:
  schemas:
    MavenBuildHandle:
      required:
      - id
      - location
      - status
      type: object
      properties:
        id:
          type: string
        location:
          type: string
          format: uri
        status:
          type: string
          enum:
          - Scheduled
          - Running
          - Success
          - Failure
          - Removed
          - NotFound
...
    UserRegister:
      required:
      - email
      - firstName
      - lastName
      - password
      - username
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        password:
          type: string
        username:
          type: string
...
    UserTokens:
      required:
      - accessToken
      - refreshToken
      - userId
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        userId:
          type: string